Move authorization to PayPalClient
