+++ src/Api/AuthorizeClientApi.php
@@ -13,39 +13,21 @@ declare(strict_types=1);
 
 namespace Sylius\PayPalPlugin\Api;
 
-use GuzzleHttp\Client;
-use Sylius\PayPalPlugin\Exception\PayPalAuthorizationException;
+use Sylius\PayPalPlugin\Client\PayPalClientInterface;
 
 final class AuthorizeClientApi implements AuthorizeClientApiInterface
 {
-    /** @var Client */
-    private $client;
+    /** @var PayPalClientInterface */
+    private $payPalClient;
 
-    /** @var string */
-    private $baseUrl;
-
-    public function __construct(Client $client, string $baseUrl)
+    public function __construct(PayPalClientInterface $payPalClient)
     {
-        $this->client = $client;
-        $this->baseUrl = $baseUrl;
+        $this->payPalClient = $payPalClient;
     }
 
     public function authorize(string $clientId, string $clientSecret): string
     {
-        $response = $this->client->request(
-            'POST',
-            $this->baseUrl . 'v1/oauth2/token',
-            [
-                'auth' => [$clientId, $clientSecret],
-                'form_params' => ['grant_type' => 'client_credentials'],
-            ]
-        );
-
-        if ($response->getStatusCode() !== 200) {
-            throw new PayPalAuthorizationException();
-        }
-
-        $content = (array) json_decode($response->getBody()->getContents(), true);
+        $content = $this->payPalClient->authorize($clientId, $clientSecret);
 
         return (string) $content['access_token'];
     }
