+++ src/Resources/config/services/api.xml
@@ -21,8 +21,7 @@
             id="Sylius\PayPalPlugin\Api\AuthorizeClientApiInterface"
             class="Sylius\PayPalPlugin\Api\AuthorizeClientApi"
         >
-            <argument type="service" id="sylius.http_client" />
-            <argument>%env(resolve:PAYPAL_API_BASE_URL)%</argument>
+            <argument type="service" id="Sylius\PayPalPlugin\Client\PayPalClientInterface" />
         </service>
 
         <service
